pyproject.toml
[tool.poetry]
name = "adventofcode2021"
version = "0.1.0"
description = ""
authors = ["Richard Liebscher <r1tschy@posteo.de>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21.4"
requests = "^2.26.0"
iterage = { git = "https://github.com/R1tschY/iterage.git" }
attrs = "*"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
poetry = "^1.1.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
